rename worker modules
